basic agent moves

+++ src/student_player/Node.java
@@ -3,9 +3,7 @@ package student_player;
 import java.util.ArrayList;
 import java.util.List;
 
-import Saboteur.SaboteurBoardState;
 import Saboteur.SaboteurMove;
-import Saboteur.cardClasses.SaboteurCard;
 
 public class Node {
     Node parent;
@@ -13,7 +11,19 @@ public class Node {
     SaboteurBoardStateClone state;
     SaboteurMove movePlayed;
     int visitCount;
-    int winScore;
+    double winScore;
+    
+//    public Node(Node parent, List<Node> childArray, int winScore, int visitCount, int turn) {
+//    	
+//    	this.parent = parent;
+//    	this.visitCount = visitCount;
+//    	this.winScore = winScore;
+//    	
+//    	if(childArray !=null) {
+//    		this.childArray = new ArrayList<Node>(childArray);
+//    	}
+//    	
+//    }
     
     public Node(SaboteurBoardStateClone state) {
         this.state = state;
@@ -64,14 +74,14 @@ public class Node {
     }
     
 
-	public int getWinScore() {
+	public double getWinScore() {
 		return winScore;
 	}
 	
 	public SaboteurMove getMovePlayed() {
 		return this.movePlayed;
 	}
-	public void setWinScore(int score) {
+	public void setWinScore(double score) {
 		this.winScore = score;
 	}
